Update WhatsApp message handling and improve cancellation support
- Added cancellation token support to message sending and WhatsApp number checking methods for better responsiveness.
- Enhanced error handling in WhatsAppMessageSender and WhatsAppUtilityController to notify users of operation cancellations.
- Updated IMessageSender and IWhatsAppService interfaces to include cancellation token parameters.
- Removed outdated whatsapp-session.rar file and updated .gitignore to reflect changes.
- Refactored various components to ensure consistent handling of user authentication and routing.

+++ src/Infrastructure/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -583,8 +583,8 @@ namespace Infrastructure.Migrations
 
                     SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
-                    b.Property<int>("ConsumedMessages")
-                        .HasColumnType("int");
+                    b.Property<long>("ConsumedMessages")
+                        .HasColumnType("bigint");
 
                     b.Property<int>("ConsumedQueues")
                         .HasColumnType("int");
@@ -604,8 +604,8 @@ namespace Infrastructure.Migrations
                     b.Property<bool>("IsDeleted")
                         .HasColumnType("bit");
 
-                    b.Property<int>("MessagesQuota")
-                        .HasColumnType("int");
+                    b.Property<long>("MessagesQuota")
+                        .HasColumnType("bigint");
 
                     b.Property<int>("ModeratorUserId")
                         .HasColumnType("int");
